devops/CI/builds_website/app/requirements.txt
flask==0.12.2
uwsgi==2.0.15
pymongo==3.7.1
boto3==1.9.22
slackclient==1.3.0
flask_dance==1.1.0
blinker==1.4
paramiko==2.3.1
redis==2.10.6
apache-libcloud==2.3.0
pycrypto==2.6.1
cloudflare==2.1.0
